requirements.txt
Cheetah==2.4.4
Django==1.5.1
## FIXME: could not find svn URL in dependency_links for this package:
Jinja2==2.7-devdev-20120723
MySQL-python==1.2.3
PAM==0.4.2
South==0.7.5
Twisted==13.0.0
Twisted-Conch==11.1.0
Twisted-Core==11.1.0
Twisted-Lore==11.1.0
Twisted-Mail==11.1.0
Twisted-Names==11.1.0
Twisted-News==11.1.0
Twisted-Runner==11.1.0
Twisted-Web==11.1.0
Twisted-Words==11.1.0
argparse==1.2.1
blackbox==0.43
carbon==0.9.10
-e git+https://github.com/graphite-project/ceres.git@d9bb2e17697cc2bc4a7c810fa116d04d9abcd3f9#egg=ceres-dev
command-not-found==0.2.44
distribute==0.6.45
django-alive==0.0.0
django-lego2==0.0.0
django-russian==0.0.0
django-tagging==0.3.1
django-template-common==0.0.0
django-yauth==0.0.0
flup==1.0.2
graphite-web==0.10.0-alpha
gunicorn==0.17.4
iotop==0.4.4
pyOpenSSL==0.12
pyasn1==0.0.11a
pycrypto==2.4.1
pyparsing==1.5.7
pyserial==2.5
python-apt==0.8.3ubuntu7
python-yenv==0.1
pytz==2013b
suds==0.4.1
txAMQP==0.6.2
virtualenv==1.7.1.2
web.py==0.34
-e git+https://github.com/graphite-project/whisper.git@3e2d7488200958ab86329ca032e17c4e97fa5dcf#egg=whisper-dev
wsgiref==0.1.2
yandex-certificator==0.0.0
ylog==0.18
zope.interface==4.0.5
